Auth and log out operational
The system is operational as far as signing up a new user, which creates an account in the database.  Right now, logs the user in upon creation, and there's a logout button at the top right of the turf screen.

When you return to the application, it keeps you logged in.

If you log out, you'll need to log in with the same handle (username) and "access key" (password) that you signed up with.  Otherwise auth will fail.

Everything at this point has been tested, restarted, and has been operational at the time of this submission.