Functionality Swap Complete
Moved the build functionality mostly to the database and back end to handle for best practices.  We built it on the front end and had a lot of trouble fixing refresh issues and log out / close application issues.

Now, we have the database holding the information for the build and everything is back to basically the functionality it was before the move.  Tested with rebuilds and everything.