File Structure Restructure
Restructured server files and inserted typescript.  Worked through those errors.  Tested in application.  Everything appears to be operational at this point - and application is successfully doing what it's supposed to be doing in this new structure.  Added nodemon to the server which is nice to not have to constantly refresh that bad boy.  I should have done that a long time ago ha.