Initial Battle Setup
Incomplete useless file for battle calculator and not connected to anything.  Just a reminder of what should be done soon.

We've got the basic model and config set up so that the model finds the right bots and such and the config helps set up the npc enum's where we'll eventually attack those types on the map.

For now, this sets up the ability to lock out a feature the user must uncover and unlock which will have information about how to do so and move the gameplay forward.