Locked until "hacked" - home > hack rig display
This added everything we needed to the database and user to determine what should be open or not (unlocked vs locked).

This is the bare bones structure we need for the battle system.  The NPC's are set up with their data.  The bots.  The player.  Everything is basically in place to execute this first "battle".