Start - iPhone 15 adjustments
Started on a 3rd gen and now have an iphone 15 simulator that shows a bunch of design flaws visually.  This is the start of working back through the application to fix these revealed flaws.  We added padding here to make it start looking better and verify we're able to connect what we're aiming for.