Battle Screen WIP: Attack Range Implemented
Initial battle nodes changed recently to include a neutral, user, and enemy control situation so we can figure out how to target each node we wanted to target.  this implementation enhances our ability to detect a node in our space when we reach it and stop the movement animation so the attack phase can begin for node control.