Damage animation & node percentage
Give a node a specific looking health bar that will go away when owned.  At this stage - I'm going to set up so that owned nodes give a small army bonus.  It will go from their starting owned nods to the neutral nodes.  Connecting network lines are where the powerups come from for the armies currently on those lines.  The go from starting to neutral, and then if neutral is owned it over takes the entire connecting line in all directions.