major refactor.  Batch 1 battle grid screen: check
We're building out single sources of truth to avoid code duplication and misunderstanding.  As a part of that process, we realized that the demo fallback battle screen was more than inefficient and needed to be disconnected.  There's an authentication piece for the battle between users that's being mocked right now, as are the battalions themselves until we're more ready to implement real battalions from the battle preparation screen.  We got the server working for the timers to avoid client manipulation.  Batch 1 internal operations for battle grid screen moved into other sources of truth where they make sense, and imported into the battle grid screen file as needed.  This will help us understand where every piece of logic should live and if it's not in there and we have rogue code, we'll be able to fix our shit.