Database backup automation & user policy
We hold the user data we now save in the database based on aggregation every 5 minutes for up to 30 days at which point it's set up to release a rolling 30 days worth of data, minimized by the aggregation input in the database so we don't cross expensive thresholds for MongoDB database space.

This is set up to record locally only so we'll need to handle production when that comes... that said - staging is handled because it's sharing a local database.  That will not be the case for the live production server and database.